fix(sync): P1: rebuild sync_lookup after provisioning in e2e/DAST setup
The "has sync_lookup been built at all" guard a facility's first sync
depends on only checks that some build has happened, not that it's
caught up with what just changed. Our upgrade-gate satisfies that guard
before provisioning ever runs, so a facility syncing immediately after
central starts can pass it while sync_lookup still predates
provisioning's data — including the facility's own row — with no error
to signal the gap. This caused DAST's authenticated facility login to
fail intermittently in CI.

Adds a standalone `rebuildSyncLookup` command (extracted from the
upgrade gate's existing rebuild-and-check logic) and calls it right
after provisioning in e2e-test-setup.sh, closing the window before any
facility attempts to sync.
